Copy permission checks from SearchResource to MessagesResource
While copying that code is certainly not ideal, I decided to hold off
on implementing a better solution until we have a better way to check
permissions in general. Better to stick to one awkward way of doing things
than inventing another slightly-better-but-not-really-good variant that's only
used in one place.